refactor(front): reorganize files per module
- This was done following Angular's best practices.
