--- frontend/src/app/models/login-request.ts
@@ -1,4 +1,0 @@
-export interface LoginRequest {
-  email: string;
-  password: string;
-}
