--- frontend/src/app/models/login-response.ts
@@ -1,7 +1,0 @@
-export interface LoginResponse {
-  isAuthSuccessful: boolean;
-  errorMessage?: string;
-  token?: string;
-  userName?: string;
-  company?: string;
-}
