+++ frontend/src/app/shared/shared.model.ts
