+++ frontend/src/app/user/user.model.ts
@@ -16,3 +16,16 @@ export interface GetUserListResponse {
   nbPages: number;
   data: UserModel[];
 }
+
+export interface LoginRequest {
+  email: string;
+  password: string;
+}
+
+export interface LoginResponse {
+  isAuthSuccessful: boolean;
+  errorMessage?: string;
+  token?: string;
+  userName?: string;
+  company?: string;
+}
